+++ flutter/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.4
+
+- Change some scoping of Providers
+  
 ## 0.0.3
 
 - Add GameServerClient